# openapi.yaml
openapi: 3.1.0
info:
  title: SentinelAPI Vulnerable Demo
  version: 1.0.0
  description: Intentionally vulnerable local-only API for authorized scanner demonstrations.
servers:
  - url: http://127.0.0.1:8000
paths:
  /health:
    get:
      responses:
        '200': { description: Service health }
  /auth/login:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200': { description: Test bearer token }
        '401': { description: Invalid test credentials }
  /orders:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current user's order summaries }
  /orders/{order_id}:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Order details (intentionally BOLA-vulnerable) }
        '401': { description: Missing or invalid token }
        '404': { description: Order missing }
  /orders/{order_id}/summary:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Owner-scoped order summary }
        '403': { description: Another user's order is forbidden }
  /orders/{order_id}/tracking:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Owner-scoped local shipping status }
        '403': { description: Another user's order is forbidden }
  /orders/{order_id}/receipt:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Owner-scoped receipt with masked payment reference }
        '403': { description: Another user's order is forbidden }
  /orders/{order_id}/items:
    get:
      security: [{ bearerAuth: [] }]
      parameters:
        - name: order_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200': { description: Owner-scoped order line items }
        '403': { description: Another user's order is forbidden }
  /profile:
    get:
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Current profile (intentionally exposes role) }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer